+++ Engine/scenes/scene_loading.h
@@ -57,7 +57,7 @@ public:
     int exec();
 
 private:
-    int _waitTimer;
+    float _waitTimer;
 
     PGE_Texture background;
     QVector<LoadingScene_misc_img > imgs;
